phase_2/requirements.txt
sentence-transformers==2.3.1
faiss-cpu==1.13.2
numpy==1.26.3
pyyaml==6.0.1
